Renaming contact generator file to contact-generator/index.ts
