+++ src/lib/engine/engine.ts
@@ -14,9 +14,9 @@ import { getEmailsForRecord } from "../model/record";
 import { Transaction } from "../model/transaction";
 import { formatMoney, formatNumber } from "../util/formatters";
 import { deriveMultiProviderDomainsSet } from "./all-free-email-providers";
-import { identifyAndFlagContactTypes } from "./contacts/contact-types";
-import { ContactGenerator } from "./contacts/generate-contacts";
-import { updateContactsBasedOnMatchResults } from "./contacts/update-contacts";
+import { ContactGenerator } from "./contact-generator";
+import { identifyAndFlagContactTypes } from "./contact-generator/contact-types";
+import { updateContactsBasedOnMatchResults } from "./contact-generator/update-contacts";
 import { DealGenerator } from "./deal-generator";
 import { LicenseGrouper } from "./license-matching/license-grouper";
 import { printSummary } from "./summary";
